requirements.txt
paho-mqtt>=1.6
python-dotenv 